requirements.txt
# React RAG Agent 系统依赖包

# 核心依赖
requests>=2.31.0
json5>=0.9.0
python-dotenv>=1.0.0

# 中文分词
jieba>=0.42.1

# 数学计算
numpy>=1.24.0

# 向量数据库
chromadb>=0.4.0

# PDF处理
PyMuPDF>=1.23.0
python-docx>=0.8.11

# 日期时间处理
python-dateutil>=2.8.2

# FastAPI 相关依赖
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# 数据库连接
mysql-connector-python>=8.2.0

# ElasticSearch客户端
elasticsearch>=8.0.0 